Move auth orchestration to app layer
