+++ src/app/providers/auth/AuthLogout.spec.tsx
@@ -74,9 +74,9 @@ vi.mock("@/pages/settings/ui/SettingsView", () => ({
 vi.mock("react-use", () => ({ useKey: vi.fn() }));
 
 import { logout } from "@/app/auth/logout";
-import { AuthProvider } from "@/app/providers/auth";
+import { AuthProvider, createAuthRuntime } from "@/app/providers/auth";
 import { RemoteReadBootstrap } from "@/app/providers/remote-read";
-import { createAuthRuntime, useAuthSession } from "@/entities/auth-session";
+import { useAuthSession } from "@/entities/auth-session";
 import { SettingsPage } from "@/pages/settings";
 import { studyStore } from "@/features/study";
