+++ src/app/providers/auth/index.ts
@@ -1,1 +1,6 @@
 export { AuthProvider } from "@/app/providers/auth/AuthProvider";
+export {
+  createAuthRuntime,
+  publishAuthenticatedUser,
+  suspendAnonymousBootstrap,
+} from "@/app/providers/auth/authController";
