+++ src/features/auth/sign-in/model/signIn.spec.ts
@@ -4,13 +4,9 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 
 const mocks = vi.hoisted(() => ({
   auth: { currentUser: null as { isAnonymous: boolean } | null },
-  publishAuthenticatedUser: vi.fn(),
 }));
 
 vi.mock("@/shared/firebase", () => ({ auth: mocks.auth }));
-vi.mock("@/entities/auth-session", () => ({
-  publishAuthenticatedUser: mocks.publishAuthenticatedUser,
-}));
 vi.mock("firebase/auth");
 
 import { loginGoogle } from "./signIn";
@@ -21,13 +17,11 @@ describe("loginGoogle", () => {
     mocks.auth.currentUser = { isAnonymous: true };
   });
 
-  it("publishes the linked user", async () => {
+  it("returns the linked user", async () => {
     const user = { uid: "uid-a" };
     vi.mocked(linkWithPopup).mockResolvedValue({ user } as never);
 
-    await loginGoogle();
-
-    expect(mocks.publishAuthenticatedUser).toHaveBeenCalledWith(user);
+    await expect(loginGoogle()).resolves.toBe(user);
   });
 
   it("recovers a credential from a Firebase linking error", async () => {
@@ -38,10 +32,9 @@ describe("loginGoogle", () => {
     vi.mocked(GoogleAuthProvider.credentialFromError).mockReturnValue(credential as never);
     vi.mocked(signInWithCredential).mockResolvedValue({ user } as never);
 
-    await loginGoogle();
+    await expect(loginGoogle()).resolves.toBe(user);
 
     expect(signInWithCredential).toHaveBeenCalledWith(mocks.auth, credential);
-    expect(mocks.publishAuthenticatedUser).toHaveBeenCalledWith(user);
   });
 
   it("rejects login without an anonymous user", async () => {
@@ -63,7 +56,6 @@ describe("loginGoogle", () => {
     vi.mocked(linkWithPopup).mockRejectedValue(error);
 
     await expect(loginGoogle()).rejects.toBe(error);
-    expect(mocks.publishAuthenticatedUser).not.toHaveBeenCalled();
   });
 
   it("preserves Firebase linking errors without a credential", async () => {
@@ -72,7 +64,6 @@ describe("loginGoogle", () => {
     vi.mocked(GoogleAuthProvider.credentialFromError).mockReturnValue(null);
 
     await expect(loginGoogle()).rejects.toBe(error);
-    expect(mocks.publishAuthenticatedUser).not.toHaveBeenCalled();
   });
 
   it("propagates credential recovery failures", async () => {
@@ -83,6 +74,5 @@ describe("loginGoogle", () => {
     vi.mocked(signInWithCredential).mockRejectedValue(recoveryError);
 
     await expect(loginGoogle()).rejects.toBe(recoveryError);
-    expect(mocks.publishAuthenticatedUser).not.toHaveBeenCalled();
   });
 });
